# back/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: MDD Application API
  version: 1.0.0
  description: |-
      This API for MDD Application allows users to access and modify their profile and subscriptions.
      This API is used to connect MDD front and back End applications.
      Some useful links:
      - [The MDD Application repository](https://github.com/ThomasBTR/Developpez-une-application-full-stack-complete)
    
  termsOfService: https://mdd-application.com/terms/
  contact:
      email: thomas.berthomier.dev@gmail.com
  license:
      name: GNU General Public License v3.0
      url: https://spdx.org/licenses/GPL-3.0-or-later.html

servers:
  - url: http://localhost:8080/api

paths:
  /users:
    get:
      tags:
        - User
      summary: Retrieve user profile
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileList'
        '401':
          description: Unauthorized access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - User
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: User profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '400':
          description: Invalid user profile data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /user/{id}:
    get:
      tags:
        - User
      summary: Retrieve user profile by ID
      parameters:
        - in: path
          name: id
          description: ID of the user to retrieve
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '400':
          description: Invalid user ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Log in user
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: User logged in successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JwtResponse'
        '401':
          description: Invalid login credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignUpRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignUpResponse'
        '400':
          description: Invalid user registration data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /logout:
    post:
      tags:
        - Authentication
      summary: Log out user
      responses:
        '204':
          description: User logged out successfully
        '401':
          description: Unauthorized access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /themes:
    get:
      tags:
        - Theme
      summary: Retrieve list of all available themes
      responses:
        '200':
          description: List of themes retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThemeListResponse'
  /subscriptions:
    post:
      tags:
          - Subscription
      summary: Subscribe to a theme
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriptionRequest'
      responses:
        '201':
          description: Subscribed to theme successfully
        '400':
          description: Invalid subscription request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /subscriptions/{theme_id}:
    delete:
      tags:
          - Subscription
      summary: Unsubscribe from a theme
      parameters:
        - in: path
          name: theme_id
          description: ID of the theme to unsubscribe from
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Unsubscribed from theme successfully
        '400':
          description: Invalid theme ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    UserDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
        password:
          type: string
    SignUpRequest:
      type: object
      properties:
          username:
            type: string
            x-field-extra-annotation: '@javax.validation.constraints.NotNull'
          email:
            type: string
            x-field-extra-annotation: '@javax.validation.constraints.NotNull'
          password:
            type: string
            x-field-extra-annotation: '@javax.validation.constraints.NotNull'
            pattern: '^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[@$!%*?&])[A-Za-z\d@$!%*?&]{8,}$'
    SignUpResponse:
        type: object
        properties:
            message:
                type: string
    UserUpdate:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
    UserProfileList:
      type: array
      items:
        $ref: '#/components/schemas/UserProfile'
    UserProfile:
      type: object
      properties:
          id:
            type: integer
            format: int64
          username:
            type: string
          email:
            type: string
          password:
            type: string
          subscriptions:
            type: array
            items:
                $ref: '#/components/schemas/ThemeListResponse'
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          x-field-extra-annotation: '@javax.validation.constraints.NotNull'
        password:
          type: string
          x-field-extra-annotation: '@javax.validation.constraints.NotNull'
          pattern: '^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[@$!%*?&])[A-Za-z\d@$!%*?&]{8,}$'
    JwtResponse:
      type: object
      properties:
          jwt:
            type: string
          type:
            type: string
          id:
            type: integer
            format: int64
          username:
            type: string
          email:
            type: string
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
         type: string
    Theme:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        users:
          type: array
          items:
            $ref: '#/components/schemas/UserThemeInfo'
    UserThemeInfo:
      type: object
      properties:
          id:
            type: integer
            format: int64
          username:
            type: string
          email:
            type: string
    ThemeListResponse:
      type: array
      items:
        $ref: '#/components/schemas/Theme'
    SubscriptionRequest:
      type: object
      properties:
        themeId:
          type: integer
          format: int64